Switch from HashSet to Dictionary for storing items in sorted folder
Keys are the short paths and the values are their filenames, since there can't be two short paths that are the same but filenames can (usually) be the same as long as in different folders.
When an item is sorted its short path and filename are added to the Dictionary, whether it goes in a year folder or unknown DT folder.