Greatly improve the Progress Bar. MoveItem() now has an optional second arg. (Check desc.)
The second argument controls if the Thread for File.Move should suspend the Thread MoveItem was called from.
I noticed before when trying AutoSort with a folder of ~3500 items, the Progress Bar would zoom to the end, but when I checked the unsorted folder in File Explorer, all the Threads that were spawned were still doing their work. So, while the GUI thought it was done, it actually wasn't. This should fix this. And, because the new arg defaults to false, when manually sorting the Threads should still work in parallel to make things faster.